+++ src/main/java/com/example/java1_2022_kub0679/Score.java
@@ -1,10 +1,10 @@
 package com.example.java1_2022_kub0679;
 
 public class Score {
-    private double coins = 750;
-    private double currentCoins;
-    private double bestScore;
-    private double currentScore;
+    private int coins = 750;
+    private int currentCoins;
+    private int bestScore;
+    private int currentScore;
 
     private int coinsMultiplier = 1;
 
@@ -13,39 +13,39 @@ public class Score {
     public Score() {
     }
 
-    public double getCoins() {
+    public int getCoins() {
         return coins;
     }
 
-    public void setCoins(double coins) {
+    public void setCoins(int coins) {
         this.coins = coins;
-        event.valueChanged("coins", (Double)coins );
+        event.valueChanged("coins", (Integer)coins );
     }
 
-    public double getCurrentCoins() {
+    public int getCurrentCoins() {
         return currentCoins;
     }
 
-    public void setCurrentCoins(double currentCoins) {
+    public void setCurrentCoins(int currentCoins) {
         this.currentCoins = currentCoins;
-        event.valueChanged("currentCoins", (Double)currentCoins );
+        event.valueChanged("currentCoins", (Integer)currentCoins );
     }
 
-    public double getBestScore() {
+    public int getBestScore() {
         return bestScore;
     }
 
-    public void setBestScore(double bestScore) {
+    public void setBestScore(int bestScore) {
         this.bestScore = bestScore;
     }
 
-    public double getCurrentScore() {
+    public int getCurrentScore() {
         return currentScore;
     }
 
-    public void setCurrentScore(double currentScore) {
+    public void setCurrentScore(int currentScore) {
         this.currentScore = currentScore;
-        event.valueChanged("score", (Double)currentScore );
+        event.valueChanged("score", (Integer)currentScore );
     }
 
     public int getCoinsMultiplier() {
@@ -55,7 +55,7 @@ public class Score {
     public void setCoinsMultiplier(int coinsMultiplier) {
         if(coinsMultiplier > 4) return;
         this.coinsMultiplier = coinsMultiplier;
-        event.valueChanged("currentCoins", (Double)currentCoins );
+        event.valueChanged("currentCoins", (Integer)currentCoins );
     }
 
     public void setEvent(GameEvent event) {
